+++ packages/@tinacms/git-client/CHANGELOG.md
@@ -13,9 +13,9 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 **Note:** Version bump only for package @tinacms/git-client
 
-## [0.4.2-alpha.1](https://github.com/tinacms/tinacms/compare/@tinacms/git-client@0.4.2-alpha.0...@tinacms/git-client@0.4.2-alpha.1) (2020-02-11)
+### Features
 
-**Note:** Version bump only for package @tinacms/git-client
+- **MediaStore:** add git backed media store ([a5b9ea4](https://github.com/tinacms/tinacms/commit/a5b9ea4))
 
 ## [0.4.2-alpha.0](https://github.com/tinacms/tinacms/compare/@tinacms/git-client@0.4.1...@tinacms/git-client@0.4.2-alpha.0) (2020-02-06)